New structure to better facilitate Dependency Injection
