+++ Finance/Data/Database/IFinanceDatabase.cs
@@ -1,0 +1,10 @@
+using System.Data.Common;
+using Finance;
+
+namespace Finance.Data.Database;
+
+public interface IFinanceDatabase
+{
+    Task<DbConnection> GetConnectionAsync();
+    Task InitializeDatabase();
+}
